Migrando views e forms de bancada para app de parlamentares
