+++ sapl/parlamentares/urls.py
@@ -18,7 +18,7 @@ from sapl.parlamentares.views import (CargoMesaCrud, ColigacaoCrud,
                                       insere_parlamentar_composicao,
                                       parlamentares_frente_selected,
                                       remove_parlamentar_composicao,
-                                      parlamentares_filiados,
+                                      parlamentares_filiados, BlocoCrud,
                                       PesquisarParlamentarView, VincularParlamentarView)
 
 from .apps import AppConfig
@@ -46,8 +46,9 @@ urlpatterns = [
 
     url(r'^sistema/coligacao/',
         include(ColigacaoCrud.get_urls() +
-                ComposicaoColigacaoCrud.get_urls())),
-
+                ComposicaoColigacaoCrud.get_urls())),    
+    url(r'^sistema/bloco/',
+        include(BlocoCrud.get_urls())),
     url(r'^sistema/frente/',
         include(FrenteCrud.get_urls())),
     url(r'^sistema/frente/atualiza-lista-parlamentares',
@@ -89,4 +90,5 @@ urlpatterns = [
 
     url(r'^mesa-diretora/remove-parlamentar-composicao/$',
         remove_parlamentar_composicao, name='remove_parlamentar_composicao'),
+    
 ]
