+++ sapl/sessao/urls.py
@@ -2,8 +2,8 @@ from django.conf.urls import include, url
 
 from sapl.sessao.views import (AdicionarVariasMateriasExpediente,
                                AdicionarVariasMateriasOrdemDia, BancadaCrud,
-                               BlocoCrud, CargoBancadaCrud,
-                               ExpedienteMateriaCrud, ExpedienteView, JustificativaAusenciaCrud,
+                               CargoBancadaCrud, ExpedienteMateriaCrud, 
+                               ExpedienteView, JustificativaAusenciaCrud,
                                OcorrenciaSessaoView, MateriaOrdemDiaCrud, OradorOrdemDiaCrud,
                                MesaView, OradorCrud,
                                OradorExpedienteCrud, PainelView,
@@ -96,8 +96,6 @@ urlpatterns = [
         include(TipoRetiradaPautaCrud.get_urls())),
     url(r'^sistema/bancada/',
         include(BancadaCrud.get_urls())),
-    url(r'^sistema/bloco/',
-        include(BlocoCrud.get_urls())),
     url(r'^sistema/cargo-bancada/',
         include(CargoBancadaCrud.get_urls())),
     url(r'^sistema/resumo-ordenacao/',
