+++ sapl/templates/menu_tabelas_auxiliares.yaml
@@ -78,7 +78,7 @@
     url: sapl.parlamentares:frente_list
     css_class: btn btn-link
   - title: {% trans 'Bloco Parlamentar' %}
-    url: sapl.sessao:bloco_list
+    url: sapl.parlamentares:bloco_list
     css_class: btn btn-link
 - title: {% trans 'Módulo Proposições' %}
   css_class: head_title
